requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.13.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
torchmetrics>=0.11.0
sentencepiece>=0.1.99
evaluate>=0.4.0
accelerate>=0.20.0
wandb>=0.15.0 